requirements.txt
PrettyTable==0.7.2

pandas==1.3.1
Django~=3.2.6
djangorestframework~=3.12.4